Update module and filter file names (.h -> .hpp)
