+++ src/filters/filter_base.hpp
