+++ src/filters/goodison_undercatch.cpp
@@ -25,7 +25,7 @@
 // Created by chris on 18/11/15.
 //
 
-#include "goodison_undercatch.h"
+#include "goodison_undercatch.hpp"
 REGISTER_FILTER_CPP(goodison_undercatch);
 
 goodison_undercatch::goodison_undercatch(config_file cfg)
