+++ src/mesh/triangulation.hpp
@@ -1,22 +1,22 @@
 // 	Copyright (C) 2011  Chris Marsh
-// 
+//
 // 	This program is free software: you can redistribute it and/or modify
 // 	it under the terms of the GNU General Public License as published by
 // 	the Free Software Foundation, either version 3 of the License, or
 // 	(at your option) any later version.
-// 
+//
 // 	This program is distributed in the hope that it will be useful,
 // 	but WITHOUT ANY WARRANTY; without even the implied warranty of
 // 	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 // 	GNU General Public License for more details.
-// 
+//
 // 	You should have received a copy of the GNU General Public License
 // 	along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 
 #pragma once
 
-#include "interpolation.h"
+#include "interpolation.hpp"
 
 
 #include "station.hpp"
@@ -697,7 +697,7 @@ private:
     //that allows us to traverse the triangulation in a deterministic order, as well as play nice with openmp
     std::vector< Delaunay::Face_handle  > _faces;
     std::vector< Delaunay::Vertex_handle > _vertexes;
-    
+
 #ifdef NOMATLAB
     //ptr to the matlab engine
     boost::shared_ptr<maw::matlab_engine> _engine;
