+++ src/modules/interp_met/Dodson_NSA_ta.cpp
@@ -21,7 +21,7 @@
 // <http://www.gnu.org/licenses/>.
 //
 
-#include "Dodson_NSA_ta.h"
+#include "Dodson_NSA_ta.hpp"
 REGISTER_MODULE_CPP(Dodson_NSA_ta);
 
 Dodson_NSA_ta::Dodson_NSA_ta(config_file cfg)
