+++ src/modules/interp_met/const_llra_ta.hpp
@@ -26,7 +26,7 @@
 #include "logger.hpp"
 #include "triangulation.hpp"
 #include "module_base.hpp"
-#include "interpolation.h"
+#include "interpolation.hpp"
 
 #include <cstdlib>
 #include <string>
