+++ src/modules/interp_met/rh_from_obs.cpp
@@ -22,7 +22,7 @@
 //
 
 
-#include "rh_from_obs.h"
+#include "rh_from_obs.hpp"
 REGISTER_MODULE_CPP(rh_from_obs);
 
 rh_from_obs::rh_from_obs(config_file cfg)
